pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
SQLAlchemy = "^1.4.35"
sqlparse = "^0.4.2"
termcolor = "^1.1.0"
Flask = "^2.1.1"
Flask-Session = "^0.4.0"
pytz = "^2022.1"
yagmail = "^0.15.277"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"